Change names and add documentation
