+++ lib/player.js
@@ -22,9 +22,9 @@ goog.require('shaka.media.PlayRateController');
 goog.require('shaka.media.Playhead');
 goog.require('shaka.media.PlayheadObserverManager');
 goog.require('shaka.media.PreferenceBasedCriteria');
+goog.require('shaka.media.QualityObserver');
 goog.require('shaka.media.RegionObserver');
 goog.require('shaka.media.RegionTimeline');
-goog.require('shaka.media.QualityTimeline');
 goog.require('shaka.media.SegmentIndex');
 goog.require('shaka.media.SrcEqualsPlayhead');
 goog.require('shaka.media.StreamingEngine');
@@ -172,6 +172,18 @@ goog.requireType('shaka.routing.Payload');
  *   'timelineregionexit'
  * @property {shaka.extern.TimelineRegionInfo} detail
  *   An object which contains a description of the region.
+ * @exportDoc
+ */
+
+/**
+ * @event shaka.Player.MediaQualityChangedEvent
+ * @description Fired when the media quality changes at the playhead.
+ * That may be caused by and adapatation change or DASH period transition.
+ * Separate events are emmited for audio and video contentTypes
+ * @property {string} type
+ *   'mediaqualitychanged'
+ * @property {shaka.extern.MediQualityInfo} detail
+ *   An object containing information about the media quality.
  * @exportDoc
  */
 
@@ -472,8 +484,8 @@ shaka.Player = class extends shaka.util.FakeEventTarget {
     /** @private {shaka.util.CmcdManager} */
     this.cmcdManager_ = null;
 
-    /** @private {shaka.media.QualityTimeline} */
-    this.qualityTimeline_ = null;
+    /** @private {shaka.media.QualityObserver} */
+    this.qualityObserver_ = null;
 
     /** @private {shaka.media.StreamingEngine} */
     this.streamingEngine_ = null;
@@ -1682,8 +1694,9 @@ shaka.Player = class extends shaka.util.FakeEventTarget {
         this.adManager_.onDashTimedMetadata(region);
       }
     });
-    this.qualityTimeline_ = new shaka.media.QualityTimeline();
-    this.qualityTimeline_.setListeners((mediaQualityInfo) => {
+    this.qualityObserver_ = new shaka.media.QualityObserver(
+        () => this.getBufferedInfo());
+    this.qualityObserver_.setListeners((mediaQualityInfo) => {
       this.onMediaQualityChange_(mediaQualityInfo);
     });
 
@@ -2651,7 +2664,7 @@ shaka.Player = class extends shaka.util.FakeEventTarget {
     goog.asserts.assert(this.manifest_, 'Must have manifest');
     goog.asserts.assert(this.regionTimeline_, 'Must have region timeline');
     goog.asserts.assert(this.video_, 'Must have video element');
-    goog.asserts.assert(this.qualityTimeline_, 'Must have quality timeline');
+    goog.asserts.assert(this.qualityObserver_, 'Must have quality timeline');
 
     // Create the region observer. This will allow us to notify the app when we
     // move in and out of timeline regions.
@@ -2675,7 +2688,7 @@ shaka.Player = class extends shaka.util.FakeEventTarget {
     // Now that we have all our observers, create a manager for them.
     const manager = new shaka.media.PlayheadObserverManager(this.video_);
     manager.manage(regionObserver);
-    manager.manage(this.qualityTimeline_);
+    manager.manage(this.qualityObserver_);
     return manager;
   }
 
@@ -2839,7 +2852,7 @@ shaka.Player = class extends shaka.util.FakeEventTarget {
       onSegmentAppended: () => this.onSegmentAppended_(),
       onInitSegmentAppended: (position, initSegment) => {
         if (initSegment.mediaQuality) {
-          this.qualityTimeline_.addMediaQualityChange(
+          this.qualityObserver_.addMediaQualityChange(
               initSegment.mediaQuality, position);
         }
       },
@@ -6449,6 +6462,7 @@ shaka.Player.EventName = {
   Loaded: 'loaded',
   Loading: 'loading',
   ManifestParsed: 'manifestparsed',
+  MediaQualityChanged: 'mediaqualitychanged',
   Metadata: 'metadata',
   OnStateChange: 'onstatechange',
   OnStateIdle: 'onstateidle',
@@ -6463,7 +6477,6 @@ shaka.Player.EventName = {
   TracksChanged: 'trackschanged',
   Unloading: 'unloading',
   VariantChanged: 'variantchanged',
-  MediaQualityChanged: 'mediaqalitychanged',
 };
